requirements.txt
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-documentai==2.3.0
boto3==1.26
amazon-textract-caller==0.0.25
amazon-textract-overlayer==0.0.10
amazon-textract-prettyprinter==0.0.16
Pillow==9.3.0
torch==1.13.0
donut-python==1.0.9
requests==2.28.1